Crate fedimint_gateway_server_db

Structs

DbKeyPrefixIter ğŸ”’

An iterator over the variants of DbKeyPrefix

FederationConfigKey ğŸ”’

FederationConfigKeyPrefix ğŸ”’

FederationConfigKeyPrefixV0 ğŸ”’

FederationConfigKeyPrefixV1 ğŸ”’

FederationConfigKeyV0 ğŸ”’

FederationConfigKeyV1 ğŸ”’

FederationConfigV0

FederationConfigV1

GatewayConfigurationKeyV0 ğŸ”’

GatewayConfigurationKeyV1

GatewayConfigurationKeyV2

GatewayConfigurationV0 ğŸ”’

GatewayConfigurationV1

GatewayConfigurationV2

GatewayPublicKey ğŸ”’

IrohKey ğŸ”’

PreimageAuthentication ğŸ”’

PreimageAuthenticationPrefix ğŸ”’

RegisteredIncomingContract

RegisteredIncomingContractKey ğŸ”’

Enums

DbKeyPrefix ğŸ”’

Traits

GatewayDbExt

GatewayDbtxNcExt

Functions

get_gatewayd_database_migrations

migrate_federation_configs ğŸ”’

migrate_to_v1 ğŸ”’

migrate_to_v2 ğŸ”’

migrate_to_v3 ğŸ”’

migrate_to_v4 ğŸ”’

migrate_to_v5 ğŸ”’

Introduced in v0.5, there is a db key clash between the FederationConfig record and the isolated databases used for each client. We must migrate the isolated databases to be behind the ClientDatabase prefix to allow the gateway to properly read the federation configs.